GitlabService: refactor
